patcher_plugin: wiring in external_script param

+++ launchers/launch_jobs_plugins/tasks/patcher_plugin.py
@@ -16,6 +16,7 @@ class CustomTask(launch_jobs.BaseTask):
         target_index = None
         target_len = None
         source = None
+        external_script = None
 
         # target_path
         try:
@@ -27,28 +28,45 @@ class CustomTask(launch_jobs.BaseTask):
         try:
             target_index = self.params["target_index"]
         except KeyError:
-            return False, "target_index is a required parameter"
+            pass # optional
 
         # target_len
         try:
             target_len = self.params["target_len"]
         except KeyError:
-            return False, "target_len is a required parameter"
+            pass # optional
 
         # source
         try:
             source = self.params["source"]
         except KeyError:
-            return False, "source is a required parameter"
+            pass # optional
 
-        return True, (target_path, target_index, target_len, source)
+        # external_script
+        try:
+            external_script = self.params["external_script"]
+        except KeyError:
+            pass # optional
+
+        if external_script is None:
+
+            if target_index is None:
+                return False, "target_index is a required parameter (when external_script is absent)"
+
+            if target_len is None:
+                return False, "target_len is a required parameter (when external_script is absent)"
+
+            if source is None:
+                return False, "source is a required parameter (when external_script is absent)"
+
+        return True, (target_path, target_index, target_len, source, external_script)
 
     def run_task(self, feedback_object, execution_name=None):
 
         v, r = self._read_params()
         if not v:
             return False, r
-        target_path, target_index, target_len, source = r
+        target_path, target_index, target_len, source, external_script = r
 
         if not os.path.exists(target_path):
             return False, "target_path [%s] does not exist" % target_path
@@ -56,26 +74,32 @@ class CustomTask(launch_jobs.BaseTask):
         if os.path.isdir(target_path):
             return False, "target_path [%s] is a folder" % target_path
 
-        contents = ""
-        with open(target_path) as f:
-            contents = f.read()
+        if external_script is not None:
 
-        target_index = int(target_index)
-        target_len = int(target_len)
+            pass # mvtodo
 
-        if target_len == 0:
-            return False, "target_len [%s] cannot be zero" % target_len
+        else:
 
-        final_pos = target_index+target_len
+            contents = ""
+            with open(target_path) as f:
+                contents = f.read()
 
-        if (final_pos) > len(contents):
-            return False, "target_index [%s] plus target_len [%s] go past the content's total length [%s]" % (target_index, target_len, len(contents))
+            target_index = int(target_index)
+            target_len = int(target_len)
 
-        new_contents = contents[0:target_index] + source + contents[target_index+target_len:]
+            if target_len == 0:
+                return False, "target_len [%s] cannot be zero" % target_len
 
-        os.remove(target_path)
+            final_pos = target_index+target_len
 
-        with open(target_path, "w") as f:
-            f.write(new_contents)
+            if (final_pos) > len(contents):
+                return False, "target_index [%s] plus target_len [%s] go past the content's total length [%s]" % (target_index, target_len, len(contents))
+
+            new_contents = contents[0:target_index] + source + contents[target_index+target_len:]
+
+            os.remove(target_path)
+
+            with open(target_path, "w") as f:
+                f.write(new_contents)
 
         return True, None
